Register Update

+++ admin/includes/view_all_posts.php
@@ -116,7 +116,7 @@
           echo "<td>$post_comment_count</td>";
           echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
           echo "<td><a href='posts.php?source=edit_posts&p_id={$post_id}'>Edit</a></td>";
-          echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+          echo "<td><a onClick=\"javascript: return confirm('Delete confirmation') \" href='posts.php?delete={$post_id}'>Delete</a></td>";
           echo "</tr>";
         }
